Tests/LibWeb/Text/expected/IntersectionObserver/implicit-root-clips-at-overflow-clip-container.txt

callback 1: ratio=0
total callbacks: 1
